TCP-DISCONNECT, TCP-ABORT, COSEM-ABORT.  All unit tests are passing.

+++ lib/DLMS-COSEM/include/StateMachine.h
@@ -64,7 +64,7 @@ namespace EPRI
         static const uint8_t TRANSITIONS[] = {\
  
 #define TRANSITION_MAP_ENTRY(State, Entry)\
-        Entry,
+        (uint8_t)(Entry),
  
 #define END_TRANSITION_MAP(RetVal, Data) \
         0 };\
